+++ pages/my-account/my-rides.vue
@@ -1,0 +1,121 @@
+<template>
+  <section class="section">
+    <div class="columns">
+      <div class="column is-6">
+        <nav>
+          <FilterDropdown
+            type="year"
+            :current="currents.year"
+            :list="selects.years"
+            @filter-results="setFilter"
+          />
+          <FilterDropdown
+            type="month"
+            :current="currents.month"
+            :list="selects.months"
+            @filter-results="setFilter"
+          />
+        </nav>
+      </div>
+      </div>
+    <div class="columns">
+      <div class="column is-5">
+        <p class="alert-message" v-if="rides === null || rides.length < 1">
+          Aucunes courses réalisées en {{ currents.month }} {{ currents.year }}
+        </p>
+        <ul v-else>
+          <li
+            v-for="ride in rides"
+          >
+            <RideCard
+              :id="ride.id"
+              :day="ride.day"
+              :hour="ride.hour"
+              :departure="ride.departure"
+              :arrival="ride.arrival"
+              :passengersCount="ride.passengersCount"
+              :luggage="ride.luggage"
+            />
+          </li>
+        </ul>
+      </div>
+    </div>
+  </section>
+</template>
+
+<script>
+import RideCard from '~/components/ride-card';
+import FilterDropdown from '~/components/elements/filter-dropdown';
+import FilterManager from '~/helpers/FilterManager';
+
+const filterManager = new FilterManager(2019);
+const selects = filterManager.getSelects();
+const currents =  filterManager.getCurrents();
+
+const mask = 'id,departure(label),arrival(label),createdAt,luggage,passengersCount';
+
+const formatData = (data) => data.map((ride) => {
+    const { day, hour } = (filterManager.formatDate(ride.createdAt));
+    return {
+      ...ride,
+      departure: ride.departure.label,
+      arrival: ride.arrival.label,
+      day,
+      hour,
+    };
+});
+
+export default {
+  data() {
+    return {
+      selects,
+      currents,
+    };
+  },
+  components: {
+    RideCard,
+    FilterDropdown,
+  },
+  async asyncData({ $api, $auth }) {
+    const { start, end } = filterManager.getFilter(currents);
+    const { data } = await $api.rides(null, mask).getRides(
+      start,
+      end,
+      {},
+      {
+        filter: {
+          userId:  $auth.user.id,
+          current: false,
+        },
+      },
+    );
+
+    return { rides: formatData(data) };
+  },
+  methods: {
+    setFilter({ key, value }) {
+      this.currents = {
+        ...this.currents,
+        [key]: value,
+      };
+      this.fetchFilter();
+    },
+    async fetchFilter() {
+        const { start, end } = filterManager.getFilter(this.currents);
+        const { data } = await this.$api.rides(null, mask).getRides(
+            start,
+            end,
+            {},
+            {
+              filter: {
+                userId: this.$auth.user.id,
+                current: false,
+              },
+            },
+        );
+
+        this.rides = formatData(data);
+      }
+    }
+};
+</script>